chore: export `Profiles.profilesPath()`
Some refactoring to facilitate this.
