+++ src/profiles/delete.ts
@@ -15,7 +15,7 @@
  */
 
 import { join } from "path"
-import { profilesPath } from "../util/cache.js"
+import { profilesPath } from "./paths.js"
 import { MadWizardOptions } from "../fe/index.js"
 
 /** Clone a profile */
