+++ src/util/cache.ts
@@ -14,10 +14,7 @@
  * limitations under the License.
  */
 
-import { join } from "path"
 import envPaths from "env-paths"
-
-import { MadWizardOptions } from "../fe/index.js"
 
 /** @return the filepath in which all persistent caches are stored */
 export function cachePath() {
@@ -27,29 +24,4 @@ export function cachePath() {
 /** @return the filepath in which all persistent data are stored */
 export function dataPath() {
   return envPaths("madwizard").data
-}
-
-/** @return the filepath in which persistent profiles are stored */
-export async function profilesPath(options: MadWizardOptions, mkdir = false) {
-  const filepath = join(options.profilesPath || process.env.MWPROFILES_PATH || dataPath(), "profiles")
-  if (mkdir) {
-    const mkdirp = await import("mkdirp").then((_) => _.default)
-    await mkdirp(filepath)
-  }
-  return filepath
-}
-
-export async function copyChoices(dstFilepath: string, options: MadWizardOptions, profile = "default") {
-  const copyFile = await import("fs").then((_) => _.copyFile)
-  const srcFilepath = join(await profilesPath(options, true), profile)
-
-  return new Promise<void>((resolve, reject) => {
-    copyFile(srcFilepath, dstFilepath, (err) => {
-      if (err) {
-        reject(err)
-      } else {
-        resolve()
-      }
-    })
-  })
 }
